requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-async-sqlalchemy==0.6.0
fastapi-pagination==0.12.14
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
loguru==0.7.2
mako==1.3.0
markupsafe==2.1.3
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.1.0
pyjwt==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
setuptools==68.2.2
sniffio==1.3.0
sqlalchemy==2.0.24
sqlalchemy-utils==0.41.1
sqlmodel==0.0.14
starlette==0.32.0
typer==0.9.0
typing-extensions==4.9.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
win32-setctime==1.1.0
boto3==1.34.14
authlib==1.3.2
itsdangerous==2.2.0
isort==5.13.2
black==24.10.0
httpx==0.27.2
requests==2.32.3
joblib==1.4.2
aiohttp==3.11.11
azure-communication-email==1.0.0
APScheduler